requirements.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1

django-cors-headers==4.3.1
gunicorn==21.2.0

celery==5.3.6
django-celery-beat==2.5.0
redis==5.0.1

drf-yasg==1.21.7
environs==11.0.0
psycopg2-binary==2.9.9

geopy==2.4.1
pillow==10.2.0
beautifulsoup4==4.12.3
lxml==5.1.0

django-filter==23.5
social-auth-app-django==5.4.0

django-storages==1.14.2
boto3==1.34.51

daphne==4.1.2
channels==4.1.0
channels-redis==4.2.0

#black==24.2.0
#flake8==7.0.0
#isort==5.13.2
#pre-commit==3.6.1